Update topic spec to any()
Hey there, thanks for the library. I'm using it for the very reason that it does support any type of value as topic (while phoenix-pubsub only supports binaries). Most often I'm using tuples as topics. This change is updating the spec for the `topic` type so that dialyzer stops complaining about my usage.

Cheers!